Use bytes instead of str when storing config to protobufs
The protobuf fields expect bytes for things like the application ID, but it
gets read from a file, so a tricky question is what type we should use where.
If we try to use bytes everywhere, that has problems because environment
variables apparently need to be str. The current approach is that we use str
everywhere, but convert to bytes when storing in protobuf format.