Do some hacks to get string types to work
The tee utility was trying to copy a binary-formatted stderr to a text-formatted
stderr, which was causing problems. Switching both to text seemed to cause
other more difficult problems, so I just made the tee utility translate binary
to text. It probably wouldn't be so hard to make it smarter later and handle all
combinations.