Change FakeFile to work with Python 3
The previous code patched the builtin file type and the builtin open function
with the FakeFile constructor, but there doesn't seem to be the same "file" in
Python 3. FakeFile was almost entirely static functions anyway, so the new code
just replaces the builtin open function (which is different from os.open, which
is also patched).