Fix some sandbox import problems
The _find_module_or_loader method was supposed to optionally return a loader,
but it but it was finding the module instead, which was crashing. I'm still not
sure why it ever worked, but removing that method call seems to fix things.

Also, builtins is a module now, so we need to whitelist.