Git rid of some of the GAE sandbox import hooks
Since the target for this project is local development and managed VMs, it
isn't super-important to maintain the GAE sandbox, and it was causing some
annoying import-related problems (in particular, the python27 urllib stub was
basically importing itself). I kept the rest of the import hooks since they
might be needed for correctness.