Don't un-import most modules on sandbox setup
The sandbox code was removing all non-approved modules from importlib after the
import hooks were in place, presumably to prevent imports from getting through
early. Since I removed most of the sandboxy (for simplicity), this step isn't
necessary anymore, and it was causing problems, so I'm removing it. In particular,
This was causing a problem in importlib._bootstrap.BYTECODE_SUFFIXES was getting
erased, without any code being run to recompute it.