Change "start_new_thread" stub to work outside a request
The code swaps out thread creation with a custom implementaiton that propagates
request state, but also assumes that the is an active request to get state for.
Either due to a Python 2/3 change or because of some detail of the sandbox
system, the previous code would ignore this override outside of a request
context, but the new code gets triggered outside a request context and crashes.
Rather than trying to stop the code path from being run, I just made it more
robust, so it now just does regular thread creation if we're not in a request.