Fix some crashes when logging
One of the logging methods was causing a crash, which was getting logged, which
caused a stack overflow error. It was writing a bytes object to a text file,
and the bytes object was created explicitly, so I just got rid of the
conversion.

Also fix a case where a string is stored as bytes in protobuf form, so I needed
to convert back.