Handle the str/bytes distinction better, mostly for protobuf code
In most cases, this is just a matter of converting to bytes when storing inside
a protobuf and converting back when reading (since protobuf code makes the most
sense on a bytes level). There were also some cases where we were unnecessarily
converting some things, which was causing problems. Also, I think HTTP payloads
need to be bytes, so we need to convert and make assertions work with bytes
there.