Fix some more Python 3 issues
When formatting protobufs, be more defensive; empty default values can still be
strings (since it's a pain to go through and change all of them), and some
values may not decode as utf8, so just convert them directly to a string.

Also fix a bug where / was being used instead of // in byte manipulation code.