Fix a crash when reading request bodies
I previously converted the request body from bytes to str on input, but later
code (webob) needs to have it still be in bytes form. So it turns out that the
right fix to webapp2 is to use BytesIO instead of StringIO, so that it properly
populates the wsgi environment.