Fix protocol buffer copying
With new-style objects, copy.copy behaves differently when working on objects
that define getstate and setstate. In particular, if getstate returns a false
object, then setstate is never called, which can result in an incomplete object.
Since all objects are new-style objects in Python 3 and protocol buffers can
serialize to an empty byte array, they were causing problems. To fix this, I
added a special truthy value to denote an empty response.