Get rid of == and != in the query DSL
The inspect module seems to assume that properties have a valid == operator. I
tried hacking it previously by trying to infer the intention from the argument
type, but it looks like that's either impossible or complicated to do in the
general case, so I'm changing the API so that you now need to say equals or
not_equal instead of using operators.

Fortunately, queries break when given True or False filters, so this will fail
fast. Unfortunately, it's kind of hard to give a good error message.